requirement.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
cryptography==37.0.2
customtkinter==4.5.2
darkdetect==0.6.0
h11==0.13.0
idna==3.3
outcome==1.2.0
Pillow==9.2.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
selenium==4.3.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.9
wsproto==1.1.0
